backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlmodel==0.0.8
pydantic==1.10.8
python-dotenv==1.0.0
sqlalchemy==1.4.41
psycopg2-binary==2.9.7
alembic==1.12.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-decouple==3.8
pytest==7.4.2
httpx==0.24.1
pytest-asyncio==0.21.1
pandas==2.1.0
openpyxl==3.1.2
reportlab==4.0.4